Add files via upload
This code is a simple menu-driven program that allows users to select various options. The program prompts the user to select either Management, User or Exit options, and based on the user's input, it either navigates the user to the relevant function or exits the program. If the user selects the User option, the program prompts the user to enter their name, email, and phone number and displays a success message. The program also includes a logout function that allows the user to log out and return to the main menu. Overall, this program demonstrates basic input and output operations in C++